+++ lib/plivo/rest_client.rb
@@ -19,6 +19,7 @@ module Plivo
     attr_reader :end_users
     attr_reader :compliance_document_types, :compliance_documents, :compliance_requirements, :compliance_applications
     attr_reader :verify_session
+    attr_reader :tollfree_verifications
 
     def initialize(auth_id = nil, auth_token = nil, proxy_options = nil, timeout = 5)
       configure_base_uri
@@ -66,6 +67,7 @@ module Plivo
       @compliance_requirements = Resources::ComplianceRequirementsInterface.new(self)
       @compliance_applications = Resources::ComplianceApplicationsInterface.new(self)
       @verify_session = Resources::SessionInterface.new(self)
+      @tollfree_verifications = Resources::TollfreeVerificationsInterface.new(self)
     end
   end
 end