+++ spec/resource_messages_spec.rb
@@ -20,7 +20,13 @@ describe 'Messages test' do
       is_domestic: message.is_domestic,
       dlt_entity_id: message.dlt_entity_id,
       dlt_template_id: message.dlt_template_id,
-      dlt_template_category: message.dlt_template_category
+      dlt_template_category: message.dlt_template_category,
+      destination_network: message.destination_network,
+      carrier_fees_rate: message.carrier_fees_rate,
+      carrier_fees: message.carrier_fees,
+      conversation_id: message.conversation_id,
+      conversation_origin: message.conversation_origin,
+      conversation_expiration_timestamp: message.conversation_expiration_timestamp
     }.to_json
   end
 
@@ -81,7 +87,11 @@ describe 'Messages test' do
     expect(response.requester_ip).to eql("192.168.1.1")
     expect(response.dlt_entity_id).to eql("1234")
     expect(response.dlt_template_id).to eql("5678")
+    expect(response.destination_network).to eql("Verizon")
     expect(response.dlt_template_category).to eql("service_implicit")
+    expect(response.conversation_id).to eql("")
+    expect(response.conversation_origin).to eql("")
+    expect(response.conversation_expiration_timestamp).to eql("")
   end
 
   it 'lists all mms media' do
@@ -134,6 +144,14 @@ describe 'Messages test' do
     expect(JSON.parse(response)['objects'][19]["dlt_entity_id"]).to eql("")
     expect(JSON.parse(response)['objects'][19]["dlt_template_id"]).to eql("")
     expect(JSON.parse(response)['objects'][19]["dlt_template_category"]).to eql("")
+
+    expect(JSON.parse(response)['objects'][1]["conversation_id"]).to eql("123")
+    expect(JSON.parse(response)['objects'][1]["conversation_origin"]).to eql("service")
+    expect(JSON.parse(response)['objects'][1]["conversation_expiration_timestamp"]).to eql("2023-08-03 23:02:00+05:30")
+
+    expect(JSON.parse(response)['objects'][19]["conversation_id"]).to eql("")
+    expect(JSON.parse(response)['objects'][19]["conversation_origin"]).to eql("")
+    expect(JSON.parse(response)['objects'][19]["conversation_expiration_timestamp"]).to eql("")
   end
 
   it 'sends a message' do