+++ spec/resource_numbers_spec.rb
@@ -22,7 +22,9 @@ describe 'Numbers test' do
       tendlc_registration_status: number.tendlc_registration_status,
       toll_free_sms_verification: number.toll_free_sms_verification,
       renewal_date: number.renewal_date,
-      cnam_lookup: number.cnam_lookup
+      cnam_lookup: number.cnam_lookup,
+      toll_free_sms_verification_id: number.toll_free_sms_verification_id,
+      toll_free_sms_verification_order_status: number.toll_free_sms_verification_order_status,
     }.to_json
   end